+++ .github/workflows/publishMavenJar4CI.yml
@@ -58,14 +58,14 @@ jobs:
       - name: Gradle Build
         working-directory: src/backend/ci
         env:
-          build_mavenRepoDeployUrl: "https://oss.sonatype.org/service/local/staging/deploy/maven2/"
+          build_mavenRepoDeployUrl: "https://ossrh-staging-api.central.sonatype.com/service/local/"
           build_mavenRepoUsername: ${{ secrets.SONATYPE_USERNAME }}
           build_mavenRepoPassword: ${{ secrets.SONATYPE_PASSWORD }}
         run: |
           echo "GITHUB_REF=${GITHUB_REF}"
           ci_version=${GITHUB_REF##*/v}
           echo "ci_version=${ci_version}"
-          ./gradlew publish -Dci_version=${ci_version} \
+          ./gradlew publish closeSonatypeStagingRepository -Dci_version=${ci_version} \
           -Psigning.keyId=${{ secrets.SIGNING_KEY_ID }} \
           -Psigning.password='${{ secrets.SIGNING_PASSWORD }}' \
           -Psigning.secretKeyRingFile=${{ github.workspace }}/secretKeyRingFile.gpg \
